Put dark brown background behind the white content background
The effect of this is revealed in browsers like Safari where you can
scroll past the top of the page before it bounces back down. Since the
top colour is dark brown, scrolling past the top to reveal a white
background is jarring. This commit makes it so a seamless dark brown
background is present instead, and moves the white background to the
content area.