Account for Twitter too-many-requests error
If the page is reloaded by the same client too many times, the twitter
API calls can reach the maximum limit for requests within a time frame.
When this happened, a Rails error was returned and the website did not
render. This change accounts for that, and displays a message in place
of the tweets instead.